repaires typename of return values in sum functions

+++ include/wipp/wippstats.h
@@ -42,11 +42,11 @@ void mean(const wipp_complex_t *buffer, size_t length, wipp_complex_t *mean);
 
 
 void sum(const double *buffer, size_t length, double *sum);
-void sum(const float *buffer, size_t length, double *sum);
-void sum(const int16_t *buffer, size_t length, double *sum);
-void sum(const int32_t *buffer, size_t length, double *sum);
-void sum(const uint16_t *buffer, size_t length, double *sum);
-void sum(const uint32_t *buffer, size_t length, double *sum);
+void sum(const float *buffer, size_t length, float *sum);
+void sum(const int16_t *buffer, size_t length, int16_t*sum);
+void sum(const int32_t *buffer, size_t length, int32_t *sum);
+void sum(const uint16_t *buffer, size_t length, uint16_t *sum);
+void sum(const uint32_t *buffer, size_t length, uint32_t *sum);
 void sum(const wipp_complex_t *buffer, size_t length, wipp_complex_t *sum);
